+++ src/resolvers/relatedWork.ts
@@ -1,26 +1,41 @@
 import { prepareObjectForLogs } from '../logger';
-import { Resolvers } from "../types";
+import { RelatedWorkSearchResults, Resolvers } from '../types';
 import { MyContext } from '../context';
 import { isAuthorized } from '../services/authService';
 import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError } from '../utils/graphQLErrors';
-import { RelatedWork } from '../models/RelatedWork';
+import { RelatedWorkSearchResult, RelatedWork, Work, WorkVersion } from '../models/RelatedWork';
 import { GraphQLError } from 'graphql';
 import { Project } from '../models/Project';
 import { hasPermissionOnProject } from '../services/projectService';
 import { Plan } from '../models/Plan';
-import { addVersion } from '../models/PlanVersion';
-import { normaliseDateTime } from "../utils/helpers";
+import { isNullOrUndefined, normaliseDateTime } from '../utils/helpers';
+import { PaginationOptionsForCursors, PaginationOptionsForOffsets, PaginationType } from '../types/general';
 
 export const resolvers: Resolvers = {
   Query: {
-    // Get all the realted works for the Project
-    async relatedWorks(_, { projectId }, context: MyContext): Promise<RelatedWork[]> {
-      const reference = 'relatedWorks resolver';
+    // Get all the related works for a plan
+    async relatedWorksByProject(
+      _,
+      { projectId, filterOptions, paginationOptions },
+      context: MyContext,
+    ): Promise<RelatedWorkSearchResults> {
+      const reference = 'relatedWorksByProject resolver';
       try {
         if (isAuthorized(context.token)) {
+          const pagOpts =
+            !isNullOrUndefined(paginationOptions) && paginationOptions.type === PaginationType.OFFSET
+              ? (paginationOptions as PaginationOptionsForOffsets)
+              : ({ ...paginationOptions, type: PaginationType.CURSOR } as PaginationOptionsForCursors);
           const project = await Project.findById(reference, context, projectId);
-          if (project && await hasPermissionOnProject(context, project)) {
-            return await RelatedWork.findByProjectId(reference, context, projectId);
+          if (project && (await hasPermissionOnProject(context, project))) {
+            return await RelatedWorkSearchResult.search(
+              reference,
+              context,
+              projectId,
+              undefined,
+              filterOptions,
+              pagOpts,
+            );
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -32,14 +47,32 @@ export const resolvers: Resolvers = {
       }
     },
 
-    async relatedWork(_, { id }, context: MyContext): Promise<RelatedWork> {
-      const reference = 'relatedWork resolver';
+    // Get all the related works for a plan
+    async relatedWorksByPlan(
+      _,
+      { planId, filterOptions, paginationOptions },
+      context: MyContext,
+    ): Promise<RelatedWorkSearchResults> {
+      const reference = 'relatedWorksByPlan resolver';
       try {
         if (isAuthorized(context.token)) {
-          const relatedWork = await RelatedWork.findById(reference, context, id);
-          const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && await hasPermissionOnProject(context, project)) {
-            return relatedWork;
+          const plan = await Plan.findById(reference, context, planId);
+          if (plan) {
+            const pagOpts =
+              !isNullOrUndefined(paginationOptions) && paginationOptions.type === PaginationType.OFFSET
+                ? (paginationOptions as PaginationOptionsForOffsets)
+                : ({ ...paginationOptions, type: PaginationType.CURSOR } as PaginationOptionsForCursors);
+            const project = await Project.findById(reference, context, plan.projectId);
+            if (project && (await hasPermissionOnProject(context, project))) {
+              return await RelatedWorkSearchResult.search(
+                reference,
+                context,
+                plan.projectId,
+                planId,
+                filterOptions,
+                pagOpts,
+              );
+            }
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -49,28 +82,58 @@ export const resolvers: Resolvers = {
         context.logger.error(prepareObjectForLogs(err), `Failure in ${reference}`);
         throw InternalServerError();
       }
-    }
+    },
   },
 
   Mutation: {
     // Add a related work to a research project
-    async addRelatedWork(_, { input }, context: MyContext): Promise<RelatedWork> {
+    async addRelatedWork(_, { input }, context: MyContext): Promise<RelatedWorkSearchResult> {
       const reference = 'addRelatedWork resolver';
       try {
         if (isAuthorized(context.token)) {
-          const project = await Project.findById(reference, context, input.projectId);
-          if (project && await hasPermissionOnProject(context, project)) {
-            const relatedWork = new RelatedWork(input);
-            const newRelatedWork = await relatedWork.create(context);
-
-            if (newRelatedWork && !newRelatedWork.hasErrors()) {
-              // Version all of the plans (if any) and sync with the DMPHub
-              const plans = await Plan.findByProjectId(reference, context, project.id);
-              for (const plan of plans) {
-                await addVersion(context, plan, reference);
+          // Check if user has permission to modify project
+          const plan = await Plan.findById(reference, context, input.planId);
+          if (plan) {
+            const project = await Project.findById(reference, context, plan.projectId);
+            if (project && (await hasPermissionOnProject(context, project))) {
+              // Fetch or create work
+              let work = await Work.findByDoi(reference, context, input.doi);
+              if (!work) {
+                work = new Work({ doi: input.doi });
+                work = await work.create(context);
               }
+
+              // Fetch or create work version
+              let workVersion = await WorkVersion.findByDoiAndHash(reference, context, input.doi, input.hash);
+              if (!workVersion) {
+                // eslint-disable-next-line @typescript-eslint/no-unused-vars
+                const { planId, doi, ...options } = input;
+                workVersion = new WorkVersion(options);
+                workVersion.workId = work.id;
+                workVersion = await workVersion.create(context, work.doi);
+              }
+              if (isNullOrUndefined(workVersion) || workVersion.hasErrors())
+              {
+                throw InternalServerError('Unable to create or find workVersion');
+              }
+
+              // Create related work
+              let relatedWork = new RelatedWork({
+                planId: input.planId,
+                workVersionId: workVersion.id,
+                status: 'ACCEPTED',
+                score: 1.0,
+                maxScore: 1.0,
+                sourceType: 'USER_ADDED',
+              });
+              relatedWork = await relatedWork.create(context);
+              if (isNullOrUndefined(relatedWork.id)) {
+                throw InternalServerError('Unable to create related work');
+              }
+
+              // Fetch and return RelatedWorkSearchResult
+              return await RelatedWorkSearchResult.findById(reference, context, relatedWork.id);
             }
-            return newRelatedWork;
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -82,64 +145,28 @@ export const resolvers: Resolvers = {
       }
     },
 
-    // Update a related work on the research project
-    async updateRelatedWork(_, { input }, context: MyContext): Promise<RelatedWork> {
-      const reference = 'updateRelatedWork resolver';
+    // Update a related work status on the research project
+    async updateRelatedWorkStatus(_, { input }, context: MyContext): Promise<RelatedWorkSearchResult> {
+      const reference = 'updateRelatedWorkStatus resolver';
       try {
         if (isAuthorized(context.token)) {
-          const relatedWork = await RelatedWork.findById(reference, context, input.relatedWorkId);
+          const relatedWork = await RelatedWork.findById(reference, context, input.id);
           if (!relatedWork) {
-            throw NotFoundError();
+            throw NotFoundError('Related work not found');
           }
 
-          const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && await hasPermissionOnProject(context, project)) {
-            const toUpdate = new RelatedWork({ ...relatedWork, ...input });
-            const updated = await toUpdate.update(context);
-
-            if(updated && !updated.hasErrors()) {
-              // Version all of the plans (if any) and sync with the DMPHub
-              const plans = await Plan.findByProjectId(reference, context, project.id);
-              for (const plan of plans) {
-                await addVersion(context, plan, reference);
-              }
-            }
-
-            return updated;
-          }
-        }
-        throw context?.token ? ForbiddenError() : AuthenticationError();
-      } catch (err) {
-        if (err instanceof GraphQLError) throw err;
-
-        context.logger.error(prepareObjectForLogs(err), `Failure in ${reference}`);
-        throw InternalServerError();
-      }
-    },
-
-    // Remove a related work from a research project
-    async removeRelatedWork(_, { id }, context: MyContext): Promise<RelatedWork> {
-      const reference = 'removeRelatedWork resolver';
-      try {
-        if (isAuthorized(context.token)) {
-          const relatedWork = await RelatedWork.findById(reference, context, id);
-          if (!relatedWork) {
-            throw NotFoundError();
+          const plan = await Plan.findById(reference, context, relatedWork.planId);
+          if (!plan) {
+            throw NotFoundError('Plan not found');
           }
 
-          const project = await Project.findById(reference, context, relatedWork.projectId);
-          if (project && await hasPermissionOnProject(context, project)) {
-            const removed = await relatedWork.delete(context);
+          const project = await Project.findById(reference, context, plan.projectId);
+          if (project && (await hasPermissionOnProject(context, project))) {
+            let toUpdate = new RelatedWork({ ...relatedWork, ...input });
+            toUpdate = await toUpdate.update(context);
 
-            if(removed && !removed.hasErrors()) {
-              // Version all of the plans (if any) and sync with the DMPHub
-              const plans = await Plan.findByProjectId(reference, context, project.id);
-              for (const plan of plans) {
-                await addVersion(context, plan, reference);
-              }
-            }
-
-            return removed;
+            // Fetch and return RelatedWorkSearchResult
+            return await RelatedWorkSearchResult.findById(reference, context, toUpdate.id);
           }
         }
         throw context?.token ? ForbiddenError() : AuthenticationError();
@@ -151,12 +178,12 @@ export const resolvers: Resolvers = {
       }
     },
   },
-  RelatedWork: {
+  RelatedWorkSearchResult: {
     created: (parent: RelatedWork) => {
       return normaliseDateTime(parent.created);
     },
     modified: (parent: RelatedWork) => {
       return normaliseDateTime(parent.modified);
-    }
-  }
+    },
+  },
 };